+++ PrimeNumbers/PrimeNumbers.h
@@ -1,7 +1,7 @@
 #ifndef PRIME_NUMBERS_DEFINED
 #define PRIME_NUMBERS_DEFINED
 
-typedef unsigned long long u128;
+typedef unsigned long long u64;
 
 #include <vector>
 #include <string>
@@ -9,14 +9,14 @@ typedef unsigned long long u128;
 class PrimeNumbers
 {
 public:
-	bool IsPrimeNumber(const u128& number);
-	std::string ShowPrimeSequence(const u128& numberOfPrimes = 0);
-	std::string OutputPrimeFactors(const u128& number);
-	void GenerateNumberOfPrimes(const u128& numberOfPrimes);
-	void GeneratePrimesUpTo(const u128& number);
+	bool IsPrimeNumber(const u64& number);
+	std::string ShowPrimeSequence(const u64& numberOfPrimes = 0);
+	std::string OutputPrimeFactors(const u64& number);
+	void GenerateNumberOfPrimes(const u64& numberOfPrimes);
+	void GeneratePrimesUpTo(const u64& number);
 private:
-	std::vector<const u128> m_PrimeNumbers;
-	bool IsDivisibleBy(const u128& total, const int division);
+	std::vector<const u64> m_PrimeNumbers;
+	bool IsDivisibleBy(const u64& total, const int division);
 	const unsigned long  m_minimumNumber = 1;
 	const unsigned int m_smallestPrime = 2;
 };
